requirements.txt
#==========для бота основные=============
aiogram~=3.20.0.post0
aiohttp  # Обязательная зависимость для aiogram
python-dotenv==1.0.0  # Для работы с .env файлами (опционально)
python-decouple~=3.8 #Для работы с файлами тоже
#==========для бота на вебхуках==========
uvicorn~=0.34.3
fastapi~=0.115.12
#============для заставки бота===========
psutil~=7.0.0